+++ package.json
@@ -9,7 +9,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "format": "prettier --write . "
   },
   "dependencies": {
     "@chakra-ui/next-js": "^2.1.3",
@@ -32,6 +33,7 @@
     "react-dom": "18.2.0",
     "react-error-overlay": "^6.0.11",
     "react-icons": "^4.9.0",
+    "react-query": "^3.39.3",
     "react-redux": "^8.0.5",
     "sass": "^1.62.1",
     "typescript": "5.0.4",
