+++ src/HOC/withAuth.tsx
@@ -1,0 +1,9 @@
+const withAuth = () => {
+
+    const authRouter = () => {
+
+    }
+
+    return authRouter; 
+
+}
