+++ src/app/home/groups.tsx
@@ -1,4 +1,4 @@
-import ContainerLayout from "@/layouts/containerLayout";
+import ContainerLayout from "@/components/templates/home/layouts/container/containerLayout";
 import { Box } from "@chakra-ui/react";
 
 const Groups = () => {
