+++ src/components/atom/home/header.tsx
@@ -1,8 +1,8 @@
 "use client";
 import { Box, Button, Flex } from "@chakra-ui/react";
-import { Logo } from "../micro/logo";
-import ContainerLayout from "@/layouts/containerLayout";
-import Buttons from "../micro/buttons";
+import { Logo } from "../icons/logo";
+import ContainerLayout from "@/components/templates/home/layouts/container/containerLayout";
+import Buttons from "../button/buttons";
 
 const Header = () => {
   return (
