+++ src/components/atom/icons/heroSvgs.tsx
