+++ src/components/templates/home/layouts/auth/authVerificationLayout.tsx
